+++ databases/alembic/versions/0ce53d526f13_rename_measure_to_condition.py
@@ -1,20 +1,22 @@
-"""Ensure Pyro5 is installed
+"""rename measure() to condition()
 
-Revision ID: 802cc65f734e
-Revises: 96ac315ec212
-Create Date: 2019-09-18 11:07:04.814951
+Revision ID: 0ce53d526f13
+Revises: 9f6bc3a1a450
+Create Date: 2019-09-22 19:44:44.105105
 
 """
+from alembic import op
+import sqlalchemy as sa
+
 
 # revision identifiers, used by Alembic.
-revision = '802cc65f734e'
-down_revision = '96ac315ec212'
+revision = '0ce53d526f13'
+down_revision = '9f6bc3a1a450'
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
-    # No database schema changes, just need to execute post-alembic code
     import sys
     import os
     sys.path.append(os.path.abspath(os.path.join(__file__, "../../../..")))