+++ databases/alembic/versions/70c828e05255_update_conditional_statements_for_new_.py
@@ -11,7 +11,7 @@ import os
 
 sys.path.append(os.path.abspath(os.path.join(__file__, "../../../..")))
 
-from databases.alembic_post import write_revision_post_alembic
+from databases.alembic_post_utils import write_revision_post_alembic
 
 
 # revision identifiers, used by Alembic.