+++ databases/alembic/versions/9f6bc3a1a450_add_line_text_to_lcd_table_and_delete_.py
@@ -20,7 +20,7 @@ def upgrade():
     # import sys
     # import os
     # sys.path.append(os.path.abspath(os.path.join(__file__, "../../../..")))
-    # from databases.alembic_post import write_revision_post_alembic
+    # from databases.alembic_post_utils import write_revision_post_alembic
     # write_revision_post_alembic(revision)
 
     with op.batch_alter_table("lcd_data") as batch_op: